+++ NEWS.md
@@ -30,6 +30,8 @@
 
 * `detect_step()` is no longer restricted to steps created in recipes (#869).
 
+* Added a new step called `step_percentile()`, that replaces the value of a variable with its percentile from the training set. (#765)
+
 * New `extract_parameter_set_dials()` and `extract_parameter_dials()` methods to extract parameter sets and single parameters from `recipe` objects.